+++ Enemies/IEnemy.cs
@@ -4,8 +4,8 @@ public interface IEnemy
 {
     public string Name { get; }
     public string ImagePath { get; }
-    public int Health { get; set; }
-    public int MaxHealth { get; }
+    public float Health { get; set; }
+    public float MaxHealth { get; }
     public int Level { get; }
     public List<string> Noise { get; }
     public Dictionary<string, float> Attacks { get; }
